Fixed some stuff for testing

+++ UTWHackathon/UTWHackathon/Robot/Robot.m
@@ -34,7 +34,7 @@
     }
 }
 
--(void)move: (int) x
+-(void)move: (NSInteger) x
 {
     _destination = x;
     if (x - self.locatorData.position.x > 0) {
